requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0
python-dotenv>=1.0.0